+++ es6/tests/angular-parser.js
@@ -1,0 +1,11 @@
+const expressions = require("angular-expressions");
+function angularParser(tag) {
+	const expr = expressions.compile(tag.replace(/’/g, "'"));
+	return {
+		get(scope) {
+			return expr(scope);
+		},
+	};
+}
+
+module.exports = angularParser;